requirements.txt
dj-database-url==0.5.0
Django==2.1.5
django-appconf==1.0.2
django-ckeditor==5.6.1
django-imagekit==4.0.2
django-js-asset==1.1.0
djangorestframework==3.9.1
pilkit==2.0
Pillow==5.4.1
python-decouple==3.1
six==1.12.0
sqlparse==0.2.4